+++ src/theme/content-page.php
@@ -1,4 +1,0 @@
-<article class="page-wrapper">
-	<h1 class="page-title"><?php the_title(); ?></h1>
-	<?php the_content(); ?>
-</article>