+++ src/theme/patterns/footer.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Title: Footer
+ * Slug: wordpressify/footer
+ * Categories: featured
+ * Keywords: Footer
+ * Block Types: core/patterns
+ */
+?>
+<!-- wp:group {"tagName":"footer","style":{"spacing":{"padding":{"top":"20px","bottom":"20px","left":"20px","right":"20px"}}},"className":"footer","layout":{"type":"default"},"metadata":{"name":"Footer"}} -->
+<footer class="wp-block-group footer" style="padding-top:20px;padding-right:20px;padding-bottom:20px;padding-left:20px"><!-- wp:paragraph {"align":"center"} -->
+<p class="has-text-align-center">Designed and developed by <a href="https://www.riangle.com"><strong>Riangle</strong></a>.</p>
+<!-- /wp:paragraph --></footer>
+<!-- /wp:group -->