+++ src/theme/patterns/header.php
@@ -1,0 +1,23 @@
+<?php
+/**
+ * Title: Header
+ * Slug: wordpressify/header
+ * Categories: featured
+ * Keywords: Header
+ * Block Types: core/patterns
+ */
+?>
+<!-- wp:group {"align":"wide","style":{"spacing":{"padding":{"bottom":"20px","top":"20px","right":"20px","left":"20px"},"blockGap":"0px"}},"className":"header","layout":{"type":"flex","flexWrap":"nowrap","justifyContent":"space-between"},"metadata":{"name":"Header"}} -->
+<div class="wp-block-group alignwide header" style="padding-top:20px;padding-right:20px;padding-bottom:20px;padding-left:20px"><!-- wp:group {"layout":{"type":"flex","flexWrap":"nowrap"}} -->
+<div class="wp-block-group"><!-- wp:image {"id":18,"width":"249px","height":"24px","scale":"cover","sizeSlug":"full","linkDestination":"custom"} -->
+<figure class="wp-block-image size-full is-resized"><a href="/"><img src="https://wordpressify.s3.eu-west-1.amazonaws.com/logo.svg" alt="" class="wp-image-18" style="object-fit:cover;width:249px;height:24px"/></a></figure>
+<!-- /wp:image -->
+
+<!-- wp:site-logo {"shouldSyncIcon":false,"className":"is-style-rounded","style":{"layout":{"selfStretch":"fixed","flexSize":"0px"}}} /--></div>
+<!-- /wp:group -->
+
+<!-- wp:group {"align":"wide","layout":{"type":"flex","flexWrap":"nowrap"}} -->
+<div class="wp-block-group alignwide"><!-- wp:navigation {"ref":25,"align":"wide","layout":{"type":"flex","setCascadingProperties":true,"justifyContent":"left"}} /--></div>
+<!-- /wp:group --></div>
+<!-- /wp:group -->
+