+++ src/theme/sidebar.php
@@ -1,5 +1,0 @@
-<div class="col-xs-12 col-lg-3">
-	<?php if (is_active_sidebar('sidebar1')) : ?>
-		<?php dynamic_sidebar('sidebar1'); ?>
-	<?php endif; ?>
-</div>